Add Offline and Manifest plugin (#687)
* Preload components/data bundles for all links regardless of browser

Gatsby core won't have SW precaching built-in in the next release so
this is necessary for a fast site. And even if a browser is precaching
bundles, pulling bundles from the SW and evaling them still adds latency
to page transitions. Doing this work ahead of time in spare moments
means that page transitions will be always be fast.

* Remove offline-plugin

The webpack offline-plugin is great but it doesn't work with content not
handled by Webpack as Gatsby is increasingly moving towards.

Also by handling this in plugins instead of core, it'll offer users more
choices about how to handle "offline" than trying to meet all needs in
core.

* Revert gatsby-link change

Overly expensive for pages with lots of links.

* This error isn't helpful. Perhaps make configurable for some APIs

* Add gatsby-plugin-offline

This adds a basic setup for handling loading a website with a service
worker generated by sw-precache. It uses the AppShell technique to load
first a simple HTML shell of the website (which would generally show the
header for most sites) which then loads the Javascript for the actual
page and renders that.

* Add gatsby-plugin-manifest — to simplify adding manifest.json to project

* Add required dependencies

* Add React as a dependency

* Remove unneeded dependency and fix function argument list

* Correct behavior for collecting pages from multiple plugins

* Correct absolute path to component

* Add plugins offfline/manifest to the GatsbyJS website

* Upgrade to latest canary + restore index.js

* Add plugins offline & manifest to gatsbygram

* Not being able to copy links from standalone display deal-breaker

* Address @jeffposnick's feedback

* Update sites to latest canary