fix(develop): query runner queue - use priority buckets for tasks instead of sorting (#12365)
## Description

This change prevents the entire task queue being sorted every time a low-priority task is added to the queue. This takes a run of `npm run develop` with 60k pages from 50 minutes to 1.5 minutes. There should be no behaviour change here, as these were already the lowest priority and will be added to the end of a FIFO queue. High priority tasks still sort, which means they will still be correctly moved to the front.

## Related Issues

Fixes one aspect of #12343